MAINT: Improved modularization of authentication
