+++ topobank/users/views.py
@@ -3,15 +3,6 @@ from django.urls import reverse
 from django.views.generic import DetailView, ListView, RedirectView, UpdateView
 from django.core.exceptions import PermissionDenied
 
-try:
-    from allauth.account.views import EmailView
-except RuntimeError:
-    # If allauth is installed but not configured
-    EmailView = None
-except ModuleNotFoundError:
-    # If allauth is not installed
-    EmailView = None
-
 from .models import User
 from .utils import are_collaborating
 
@@ -51,7 +42,6 @@ class UserRedirectView(LoginRequiredMixin, RedirectView):
 
 
 class UserUpdateView(LoginRequiredMixin, UpdateView):
-
     fields = ["name"]
 
     # we already imported User in the view code above, remember?
@@ -83,27 +73,6 @@ class UserUpdateView(LoginRequiredMixin, UpdateView):
             }
         ]
         return context
-
-
-if EmailView is not None:
-    class TabbedEmailView(EmailView):
-        def get_context_data(self, **kwargs):
-            context = super().get_context_data(**kwargs)
-            context['extra_tabs'] = [
-                {
-                    'title': f"User Profile",
-                    'icon': "user",
-                    'href': reverse('users:detail', kwargs=dict(username=self.request.user.username)),
-                    'active': False
-                },
-                {
-                    'title': f"Edit E-mail Addresses",
-                    'icon': "edit",
-                    'href': self.request.path,
-                    'active': True
-                }
-            ]
-            return context
 
 
 class UserListView(LoginRequiredMixin, ListView):
